requirements.txt
datasets
transformers
torch
jupyter
matplotlib
pandas
python-dotenv
google-genai
